+++ SinZsEventTester/docs/Changelog.md
@@ -3,7 +3,7 @@
 
 ### Version 0.1.7
 * Added dialogue checker.
-* Item Spawn Query checker now checks for error items.
+* Item Spawn Query checker now checks for error items. Tighted rules on item spawn query replacements.
 
 ### Version 0.1.6
 * Added ability to fast forward.